Add hosted files
In preparation for actually using the design-system repo as a submodule
in the RPI backend, we have to (temporarily) commit also non-design
related files that are hosted through the express `/public` folder
currently.

The goal today is to use the design system repository immediately, while
for the future we can find the appropriate place to hoste these files.