mv index.html test.html
As we're now using this repository as a submodule in the RPI backend, we
can't host an index.html file as the express.js server would serve it to
all users as a static file. Hence we're changing the instructions on how
to work with the rugpullindex/design-system such that there's a
test.html file.